实现了购物车移除商品功能，计算总价格功能，优化前端设计

+++ javawork/src/cn/edu/swu/zl/goods/SellGoodsService.java
@@ -34,7 +34,7 @@ public class SellGoodsService {
 	    	sGood.setName(rs.getString("name"));
 	    	sGood.setImgUrl(rs.getString("imgUrl"));
 	    	sGood.setDscp(rs.getString("dscp"));
-	    	sGood.setPrice(rs.getInt("price"));
+	    	sGood.setPrice(rs.getFloat("price"));
 	    	sGood.setCount(rs.getInt("count"));
 	    	sGoods.add(sGood);
 	    }
